+++ src/web/webfrontend/src/home.js
@@ -1,7 +1,8 @@
 // Home page (Chat) functionality
 document.addEventListener('DOMContentLoaded', () => {
     const isLocalhost = window.location.hostname === 'localhost' || window.location.hostname === '127.0.0.1';
-    const API_BASE = isLocalhost ? 'http://localhost:80' : '';
+    // Use empty string for same-origin (webserver proxies API calls to gateway)
+    const API_BASE = '';
     const messageInput = document.getElementById('messageInput');
     const sendBtn = document.getElementById('sendBtn');
     const messagesContainer = document.getElementById('messages');
