+++ src/web/webfrontend/src/login.js
@@ -1,18 +1,16 @@
-// Register functionality
+// Login functionality
 document.addEventListener('DOMContentLoaded', () => {
     const isLocalhost = window.location.hostname === 'localhost' || window.location.hostname === '127.0.0.1';
-    const API_BASE = isLocalhost ? 'http://localhost:80' : '';
-    const registerForm = document.getElementById('registerForm');
-    const nameInput = document.getElementById('name');
+    // Use empty string for same-origin (webserver proxies API calls to gateway)
+    const API_BASE = '';
+    const loginForm = document.getElementById('loginForm');
     const emailInput = document.getElementById('email');
     const passwordInput = document.getElementById('password');
-    const confirmPasswordInput = document.getElementById('confirmPassword');
     const submitBtn = document.getElementById('submitBtn');
     const messageDiv = document.getElementById('message');
     const passwordToggle = document.getElementById('passwordToggle');
-    const confirmPasswordToggle = document.getElementById('confirmPasswordToggle');
     
-    // Password toggle for password field
+    // Password toggle
     if (passwordToggle) {
         passwordToggle.addEventListener('click', () => {
             const type = passwordInput.type === 'password' ? 'text' : 'password';
@@ -21,75 +19,51 @@ document.addEventListener('DOMContentLoaded', () => {
         });
     }
     
-    // Password toggle for confirm password field
-    if (confirmPasswordToggle) {
-        confirmPasswordToggle.addEventListener('click', () => {
-            const type = confirmPasswordInput.type === 'password' ? 'text' : 'password';
-            confirmPasswordInput.type = type;
-            confirmPasswordToggle.textContent = type === 'password' ? '👁️' : '🙈';
-        });
-    }
-    
     // Form validation
     const validateForm = () => {
-        const name = nameInput.value.trim();
         const email = emailInput.value.trim();
-        const password = passwordInput.value;
-        const confirmPassword = confirmPasswordInput.value;
+        const password = passwordInput.value.trim();
         
-        if (name === '' || email === '' || password === '' || confirmPassword === '') {
-            return { valid: false, message: 'All fields are required' };
+        if (email === '' || password === '') {
+            return false;
         }
         
         // Basic email validation
         const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
         if (!emailRegex.test(email)) {
-            return { valid: false, message: 'Please enter a valid email' };
+            return false;
         }
         
-        // Password length check
-        if (password.length < 8) {
-            return { valid: false, message: 'Password must be at least 8 characters' };
-        }
-        
-        // Password match check
-        if (password !== confirmPassword) {
-            return { valid: false, message: 'Passwords do not match' };
-        }
-        
-        return { valid: true };
+        return true;
     };
     
     // Enable/disable submit button based on validation
-    [nameInput, emailInput, passwordInput, confirmPasswordInput].forEach(input => {
+    [emailInput, passwordInput].forEach(input => {
         input.addEventListener('input', () => {
-            const validation = validateForm();
-            submitBtn.disabled = !validation.valid;
+            submitBtn.disabled = !validateForm();
         });
     });
     
     // Handle form submission
-    registerForm.addEventListener('submit', async (e) => {
+    loginForm.addEventListener('submit', async (e) => {
         e.preventDefault();
         
-        const validation = validateForm();
-        if (!validation.valid) {
-            showMessage(validation.message, 'error');
+        if (!validateForm()) {
+            showMessage('Please fill in all fields correctly', 'error');
             return;
         }
         
         submitBtn.disabled = true;
-        submitBtn.textContent = 'Creating account...';
+        submitBtn.textContent = 'Signing in...';
         
         try {
-            const response = await fetch(`${API_BASE}/api/auth/register`, {
+            const response = await fetch(`${API_BASE}/api/auth/login`, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
                 },
                 credentials: 'include',
                 body: JSON.stringify({
-                    username: nameInput.value.trim(),
                     email: emailInput.value.trim(),
                     password: passwordInput.value,
                 }),
@@ -98,26 +72,25 @@ document.addEventListener('DOMContentLoaded', () => {
             const data = await response.json();
             
             if (response.ok) {
-                showMessage('Account created successfully! Signing you in...', 'success');
+                showMessage('Login successful! Redirecting...', 'success');
                 // Store real session info from gateway response
                 localStorage.setItem('sessionId', data.session_id || ('session_' + Date.now()));
                 localStorage.setItem('userId', data.user_id || emailInput.value.trim());
-                localStorage.setItem('username', data.username || nameInput.value.trim());
-                
+                localStorage.setItem('username', data.username || emailInput.value.trim().split('@')[0]);
                 // Redirect to home page
                 setTimeout(() => {
                     window.location.href = '/home.html';
-                }, 2000);
+                }, 1000);
             } else {
-                showMessage(data.detail || data.message || 'Registration failed. Please try again.', 'error');
+                showMessage(data.detail || data.message || 'Login failed. Please try again.', 'error');
                 submitBtn.disabled = false;
-                submitBtn.textContent = 'Create Account';
+                submitBtn.textContent = 'Sign In';
             }
         } catch (error) {
-            console.error('Registration error:', error);
+            console.error('Login error:', error);
             showMessage('An error occurred. Please try again.', 'error');
             submitBtn.disabled = false;
-            submitBtn.textContent = 'Create Account';
+            submitBtn.textContent = 'Sign In';
         }
     });
