+++ src/web/webfrontend/src/register.js
@@ -1,15 +1,19 @@
-// Login functionality
+// Register functionality
 document.addEventListener('DOMContentLoaded', () => {
     const isLocalhost = window.location.hostname === 'localhost' || window.location.hostname === '127.0.0.1';
-    const API_BASE = isLocalhost ? 'http://localhost:80' : '';
-    const loginForm = document.getElementById('loginForm');
+    // Use empty string for same-origin (webserver proxies API calls to gateway)
+    const API_BASE = '';
+    const registerForm = document.getElementById('registerForm');
+    const nameInput = document.getElementById('name');
     const emailInput = document.getElementById('email');
     const passwordInput = document.getElementById('password');
+    const confirmPasswordInput = document.getElementById('confirmPassword');
     const submitBtn = document.getElementById('submitBtn');
     const messageDiv = document.getElementById('message');
     const passwordToggle = document.getElementById('passwordToggle');
+    const confirmPasswordToggle = document.getElementById('confirmPasswordToggle');
     
-    // Password toggle
+    // Password toggle for password field
     if (passwordToggle) {
         passwordToggle.addEventListener('click', () => {
             const type = passwordInput.type === 'password' ? 'text' : 'password';
@@ -18,51 +22,75 @@ document.addEventListener('DOMContentLoaded', () => {
         });
     }
     
+    // Password toggle for confirm password field
+    if (confirmPasswordToggle) {
+        confirmPasswordToggle.addEventListener('click', () => {
+            const type = confirmPasswordInput.type === 'password' ? 'text' : 'password';
+            confirmPasswordInput.type = type;
+            confirmPasswordToggle.textContent = type === 'password' ? '👁️' : '🙈';
+        });
+    }
+    
     // Form validation
     const validateForm = () => {
+        const name = nameInput.value.trim();
         const email = emailInput.value.trim();
-        const password = passwordInput.value.trim();
+        const password = passwordInput.value;
+        const confirmPassword = confirmPasswordInput.value;
         
-        if (email === '' || password === '') {
-            return false;
+        if (name === '' || email === '' || password === '' || confirmPassword === '') {
+            return { valid: false, message: 'All fields are required' };
         }
         
         // Basic email validation
         const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
         if (!emailRegex.test(email)) {
-            return false;
+            return { valid: false, message: 'Please enter a valid email' };
         }
         
-        return true;
+        // Password length check
+        if (password.length < 8) {
+            return { valid: false, message: 'Password must be at least 8 characters' };
+        }
+        
+        // Password match check
+        if (password !== confirmPassword) {
+            return { valid: false, message: 'Passwords do not match' };
+        }
+        
+        return { valid: true };
     };
     
     // Enable/disable submit button based on validation
-    [emailInput, passwordInput].forEach(input => {
+    [nameInput, emailInput, passwordInput, confirmPasswordInput].forEach(input => {
         input.addEventListener('input', () => {
-            submitBtn.disabled = !validateForm();
+            const validation = validateForm();
+            submitBtn.disabled = !validation.valid;
         });
     });
     
     // Handle form submission
-    loginForm.addEventListener('submit', async (e) => {
+    registerForm.addEventListener('submit', async (e) => {
         e.preventDefault();
         
-        if (!validateForm()) {
-            showMessage('Please fill in all fields correctly', 'error');
+        const validation = validateForm();
+        if (!validation.valid) {
+            showMessage(validation.message, 'error');
             return;
         }
         
         submitBtn.disabled = true;
-        submitBtn.textContent = 'Signing in...';
+        submitBtn.textContent = 'Creating account...';
         
         try {
-            const response = await fetch(`${API_BASE}/api/auth/login`, {
+            const response = await fetch(`${API_BASE}/api/auth/register`, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
                 },
                 credentials: 'include',
                 body: JSON.stringify({
+                    username: nameInput.value.trim(),
                     email: emailInput.value.trim(),
                     password: passwordInput.value,
                 }),
@@ -71,25 +99,26 @@ document.addEventListener('DOMContentLoaded', () => {
             const data = await response.json();
             
             if (response.ok) {
-                showMessage('Login successful! Redirecting...', 'success');
+                showMessage('Account created successfully! Signing you in...', 'success');
                 // Store real session info from gateway response
                 localStorage.setItem('sessionId', data.session_id || ('session_' + Date.now()));
                 localStorage.setItem('userId', data.user_id || emailInput.value.trim());
-                localStorage.setItem('username', data.username || emailInput.value.trim().split('@')[0]);
+                localStorage.setItem('username', data.username || nameInput.value.trim());
+                
                 // Redirect to home page
                 setTimeout(() => {
                     window.location.href = '/home.html';
-                }, 1000);
+                }, 2000);
             } else {
-                showMessage(data.detail || data.message || 'Login failed. Please try again.', 'error');
+                showMessage(data.detail || data.message || 'Registration failed. Please try again.', 'error');
                 submitBtn.disabled = false;
-                submitBtn.textContent = 'Sign In';
+                submitBtn.textContent = 'Create Account';
             }
         } catch (error) {
-            console.error('Login error:', error);
+            console.error('Registration error:', error);
             showMessage('An error occurred. Please try again.', 'error');
             submitBtn.disabled = false;
-            submitBtn.textContent = 'Sign In';
+            submitBtn.textContent = 'Create Account';
         }
     });
